requirements.txt
torch
rdkit-pypi
transformers
pandas
